Add glyph_cache.rs — VecAtlas-backed LRU glyph cache
Implements VecAtlas (fixed-block storage) with LRU bookkeeping and key→slot mapping.
Provides GlyphCache and GlyphCacheItem to select an appropriate cache by glyph bitmap size and insert/retrieve rasterized data.

others:

Added: Bitmap struct in renderer.rs with width, height, pixels, and an accumulate method that handles saturated addition.
Refactored: cpu_renderer.rs rewritten to integrate a glyph_cache module;
API changes: Simplified CpuRenderer type and method signatures — render_layout now returns a Bitmap; accessors for the cache were added.
Fix: Cast width/height to u32 in debug_image.rs ImageBuffer::from_raw(...).
Updated: debug_renderer.rs to use Bitmap and call accumulate for pixel updates, simplifying bounds handling.
Dependencies: Added quick_cache = "0.6.18" to Cargo.toml and moved image = "0.25.8" into [dev-dependencies].
Other: Minor formatting/signature change in layout.rs.

+++ examples/debug_image.rs
@@ -114,7 +114,7 @@ fn main() {
     }
 
     let img_buffer: ImageBuffer<Luma<u8>, Vec<u8>> =
-        ImageBuffer::from_raw(bitmap.width, bitmap.height, bitmap.pixels)
+        ImageBuffer::from_raw(bitmap.width as u32, bitmap.height as u32, bitmap.pixels)
             .expect("bitmap dimensions must match pixel buffer length");
 
     img_buffer
